# Backendbooking/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Space Booking API
  version: 1.0.0
  description: |
    Kontrak 50 operasi UKK Paket B. Semua data bisnis tenant-scoped.
    x-maker-key dan x-app-key adalah alias; jika keduanya hadir nilainya harus sama.
    Kontrak ini sengaja memperketat upload menjadi bearer-only, melindungi maker/list,
    menggunakan QR opaque tanpa app key, dan mempertahankan body array collection dengan
    metadata pagination pada X-Page/X-Per-Page/X-Total-Count. Lihat api-contract.md untuk conflict log.
servers:
  - url: https://learn.smktelkom-mlg.sch.id/coworking
    description: Base URL contoh sumber UKK
  - url: http://localhost:3000
    description: Development lokal
...
paths:
...
  /api/diskon/active:
    get:
      tags: [Discounts]
      operationId: listActiveDiscounts
      summary: Daftar promo aktif
      security:
        - makerKey: []
        - appKey: []
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Promo aktif
          headers:
            X-Page: { $ref: '#/components/headers/XPage' }
            X-Per-Page: { $ref: '#/components/headers/XPerPage' }
            X-Total-Count: { $ref: '#/components/headers/XTotalCount' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscountListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/diskon/check:
    post:
      tags: [Discounts]
      operationId: checkDiscountCode
      summary: Validasi kode promo
      security:
        - makerKey: []
        - appKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CheckPromoDto' }
      responses:
        '200':
          description: Promo valid
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscountCheckResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/diskon/{id}:
    get:
      tags: [Discounts]
      operationId: getPublicDiscountById
      summary: Detail diskon
      security:
        - makerKey: []
        - appKey: []
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200':
          description: Detail diskon
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscountResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/reservasi:
    post:
      tags: [Reservations]
      operationId: createReservation
      summary: Membuat reservasi member secara atomik
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string, minLength: 8, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateReservationDto' }
      responses:
        '201':
          description: Reservasi dibuat
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReservationResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/reservasi/my:
    get:
      tags: [Reservations]
      operationId: listMyReservations
      summary: Daftar reservasi member sendiri
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Reservasi sendiri
          headers:
            X-Page: { $ref: '#/components/headers/XPage' }
            X-Per-Page: { $ref: '#/components/headers/XPerPage' }
            X-Total-Count: { $ref: '#/components/headers/XTotalCount' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReservationSummaryListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/reservasi/my/history:
    get:
      tags: [Reservations]
      operationId: getMyReservationHistory
      summary: Histori reservasi member per bulan
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - { $ref: '#/components/parameters/Month' }
        - { $ref: '#/components/parameters/Year' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Histori bulanan; data.items tetap array source-compatible
          headers:
            X-Page: { $ref: '#/components/headers/XPage' }
            X-Per-Page: { $ref: '#/components/headers/XPerPage' }
            X-Total-Count: { $ref: '#/components/headers/XTotalCount' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReservationHistoryResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/reservasi/{id}/e-ticket:
    get:
      tags: [Reservations]
      operationId: getReservationETicket
      summary: E-ticket reservasi dan inspeksi tiket sebelum check-in
      description: QR memakai token opaque/signed dan tidak memuat app key seperti contoh sumber. Jalur scan/validate wajib menginspeksi e-ticket ini dalam scope admin/member yang sah, lalu admin memanggil EP-044 menggunakan id reservasi yang cocok; endpoint QR verify khusus tetap optional.
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200':
          description: E-ticket
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ETicketResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/reservasi/{id}:
    get:
      tags: [Reservations]
      operationId: getReservationById
      summary: Detail reservasi untuk pemilik atau admin owner
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200':
          description: Detail reservasi
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReservationDetailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/reservasi/{id}/cancel:
    patch:
      tags: [Reservations]
      operationId: cancelMyReservation
      summary: Membatalkan reservasi member
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200':
          description: Reservasi dibatalkan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReservationStatusResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/admin/diskon:
    get:
      tags: [Admin Discounts]
      operationId: listAdminDiscounts
      summary: Daftar semua promo admin
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Daftar diskon
          headers:
            X-Page: { $ref: '#/components/headers/XPage' }
            X-Per-Page: { $ref: '#/components/headers/XPerPage' }
            X-Total-Count: { $ref: '#/components/headers/XTotalCount' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscountListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [Admin Discounts]
      operationId: createAdminDiscount
      summary: Tambah promo
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDiscountDto' }
      responses:
        '201':
          description: Diskon dibuat
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscountResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/admin/diskon/{id}:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    get:
      tags: [Admin Discounts]
      operationId: getAdminDiscountById
      summary: Detail promo admin
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      responses:
        '200':
          description: Detail diskon
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscountResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      tags: [Admin Discounts]
      operationId: updateAdminDiscount
      summary: Update promo admin
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateDiscountDto' }
      responses:
        '200':
          description: Diskon diperbarui
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscountResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [Admin Discounts]
      operationId: deleteAdminDiscount
      summary: Arsipkan/hapus promo tanpa merusak histori
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      responses:
        '200':
          description: Diskon diarsipkan atau dihapus
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeleteResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT App Maker, member, admin_space, atau Guru/Penguji sesuai endpoint.
    makerKey:
      type: apiKey
      in: header
      name: x-maker-key
      description: Tenant key utama.
    appKey:
      type: apiKey
      in: header
      name: x-app-key
      description: Alias kompatibilitas untuk x-maker-key; bila keduanya ada harus sama.
  headers:
    XPage:
      description: Nomor halaman efektif (1-based).
      schema: { type: integer, minimum: 1 }
    XPerPage:
      description: Batas efektif item per halaman.
      schema: { type: integer, minimum: 1, maximum: 100 }
    XTotalCount:
      description: Jumlah total item yang cocok sebelum pagination.
      schema: { type: integer, minimum: 0 }
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/PositiveId' }
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Month:
      name: month
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 12 }
    Year:
      name: year
      in: query
      required: false
      schema: { type: integer, minimum: 2000, maximum: 2100 }
  responses:
    BadRequest:
      description: Input/header tidak valid atau aturan bisnis dasar gagal
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { status: false, statusCode: 400, message: Permintaan tidak valid, error: Bad Request, timestamp: '2026-08-27T08:00:00.000Z' }
    Unauthorized:
      description: Credential hilang atau invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Role/ownership/state tidak mengizinkan
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource tidak ada atau disamarkan karena scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Duplicate, overlap, stale write, atau transisi ilegal
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    TooManyRequests:
      description: Rate limit terlampaui
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
          description: Detik sebelum mencoba kembali
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalError:
      description: Kegagalan internal tanpa stack/SQL/secret
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    PositiveId:
      type: integer
      format: int64
      minimum: 1
      example: 1
    Money:
      type: integer
      format: int64
      minimum: 0
      description: Nominal integer Rupiah (IDR)
      example: 25000
    DateOnly:
      type: string
      format: date
      example: '2026-08-30'
    TimeHHmm:
      type: string
      pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
      example: '09:00'
    SpaceType:
      type: string
      enum: [desk, meeting_room, private_office]
...
    ReservationStatus:
      type: string
      enum: [belum_dikonfirm, disetujui, aktif, selesai, dibatalkan]
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [status, statusCode, message, error, timestamp]
      properties:
        status: { type: boolean, enum: [false] }
        statusCode: { type: integer, enum: [400, 401, 403, 404, 409, 413, 429, 500] }
        message: { type: string }
        error: { type: string }
        timestamp: { type: string, format: date-time }
...
    SuccessBase:
      type: object
      required: [status, statusCode, message, timestamp]
      properties:
        status: { type: boolean, enum: [true] }
        statusCode: { type: integer, enum: [200, 201] }
        message: { type: string }
        timestamp: { type: string, format: date-time }
...
    Discount:
      type: object
      additionalProperties: false
      required: [id, nama_diskon, persentase_diskon, tanggal_awal, tanggal_akhir]
      properties:
        id: { $ref: '#/components/schemas/PositiveId' }
        nama_diskon: { type: string, pattern: '^[A-Z0-9_-]+$', example: DISKONHEMAT20 }
        persentase_diskon: { type: integer, minimum: 1, maximum: 100 }
        tanggal_awal: { type: string, format: date-time }
        tanggal_akhir: { type: string, format: date-time }
    DiscountResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/Discount' }
    DiscountListResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data:
              type: array
              maxItems: 100
              items: { $ref: '#/components/schemas/Discount' }
    CheckPromoDto:
      type: object
      additionalProperties: false
      required: [nama_diskon]
      properties:
        nama_diskon: { type: string, minLength: 1, maxLength: 100, example: DISKONHEMAT20 }
    DiscountCheck:
      allOf:
        - { $ref: '#/components/schemas/Discount' }
        - type: object
          required: [is_active]
          properties:
            is_active: { type: boolean, enum: [true] }
    DiscountCheckResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/DiscountCheck' }
    CreateReservationDto:
      type: object
      additionalProperties: false
      required: [id_space, tanggal_reservasi, jam_mulai, durasi_jam]
      properties:
        id_space: { $ref: '#/components/schemas/PositiveId' }
        tanggal_reservasi: { $ref: '#/components/schemas/DateOnly' }
        jam_mulai: { $ref: '#/components/schemas/TimeHHmm' }
        durasi_jam: { type: integer, minimum: 1, example: 3 }
        id_diskon: { type: integer, format: int64, minimum: 1, nullable: true }
        kode_promo: { type: string, minLength: 1, maxLength: 100, nullable: true, example: DISKONHEMAT20 }
    Reservation:
      type: object
      additionalProperties: false
      required: [id, kode_booking, id_member, id_space, tanggal_reservasi, jam_mulai, jam_selesai, durasi_jam, harga_per_jam, total_harga_awal, potongan_diskon, total_bayar, status]
      properties:
        id: { $ref: '#/components/schemas/PositiveId' }
        kode_booking: { type: string, example: BOOK-20260830-0012 }
        id_member: { $ref: '#/components/schemas/PositiveId' }
        id_space: { $ref: '#/components/schemas/PositiveId' }
        id_diskon: { type: integer, format: int64, minimum: 1, nullable: true }
        tanggal_reservasi: { $ref: '#/components/schemas/DateOnly' }
        jam_mulai: { $ref: '#/components/schemas/TimeHHmm' }
        jam_selesai: { $ref: '#/components/schemas/TimeHHmm' }
        durasi_jam: { type: integer, minimum: 1 }
        harga_per_jam: { $ref: '#/components/schemas/Money' }
        total_harga_awal: { $ref: '#/components/schemas/Money' }
        potongan_diskon: { $ref: '#/components/schemas/Money' }
        total_bayar: { $ref: '#/components/schemas/Money' }
        status: { $ref: '#/components/schemas/ReservationStatus' }
        created_at: { type: string, format: date-time }
    ReservationResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/Reservation' }
    SpaceSummary:
      type: object
      additionalProperties: false
      required: [id, nama_space, tipe]
      properties:
        id: { $ref: '#/components/schemas/PositiveId' }
        nama_space: { type: string }
        tipe: { $ref: '#/components/schemas/SpaceType' }
        harga_per_jam: { $ref: '#/components/schemas/Money' }
    ReservationSummary:
      type: object
      additionalProperties: false
      required: [id, kode_booking, tanggal_reservasi, jam_mulai, jam_selesai, durasi_jam, total_bayar, status, space]
      properties:
        id: { $ref: '#/components/schemas/PositiveId' }
        kode_booking: { type: string }
        tanggal_reservasi: { $ref: '#/components/schemas/DateOnly' }
        jam_mulai: { $ref: '#/components/schemas/TimeHHmm' }
        jam_selesai: { $ref: '#/components/schemas/TimeHHmm' }
        durasi_jam: { type: integer, minimum: 1 }
        total_bayar: { $ref: '#/components/schemas/Money' }
        status: { $ref: '#/components/schemas/ReservationStatus' }
        space: { $ref: '#/components/schemas/SpaceSummary' }
    ReservationSummaryListResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data:
              type: array
              maxItems: 100
              items: { $ref: '#/components/schemas/ReservationSummary' }
    HistoryItem:
      type: object
      additionalProperties: false
      required: [id, kode_booking, tanggal_reservasi, jam_mulai, jam_selesai, durasi_jam, total_bayar, status, space_name]
      properties:
        id: { $ref: '#/components/schemas/PositiveId' }
        kode_booking: { type: string }
        tanggal_reservasi: { $ref: '#/components/schemas/DateOnly' }
        jam_mulai: { $ref: '#/components/schemas/TimeHHmm' }
        jam_selesai: { $ref: '#/components/schemas/TimeHHmm' }
        durasi_jam: { type: integer, minimum: 1 }
        total_bayar: { $ref: '#/components/schemas/Money' }
        status: { $ref: '#/components/schemas/ReservationStatus' }
        space_name: { type: string }
    ReservationHistory:
      type: object
      additionalProperties: false
      required: [month, year, total_reservasi, total_pengeluaran, items]
      properties:
        month: { type: integer, minimum: 1, maximum: 12 }
        year: { type: integer, minimum: 2000, maximum: 2100 }
        total_reservasi: { type: integer, minimum: 0 }
        total_pengeluaran: { $ref: '#/components/schemas/Money' }
        items:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/HistoryItem' }
    ReservationHistoryResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/ReservationHistory' }
    ETicket:
      type: object
      additionalProperties: false
      required: [e_ticket_number, kode_booking, coworking_space, member, space, jadwal, rincian_pembayaran, status_reservasi, qr_code_payload]
      properties:
        e_ticket_number: { type: string, example: TICKET-MOKLET-20260830-0012 }
        kode_booking: { type: string }
        coworking_space:
          type: object
          additionalProperties: false
          required: [nama, telepon]
          properties:
            nama: { type: string }
            telepon: { type: string }
        member:
          type: object
          additionalProperties: false
          required: [nama, instansi, telp]
          properties:
            nama: { type: string }
            instansi: { type: string }
            telp: { type: string }
        space:
          type: object
          additionalProperties: false
          required: [nama, tipe, harga_per_jam]
          properties:
            nama: { type: string }
            tipe: { type: string }
            harga_per_jam: { $ref: '#/components/schemas/Money' }
        jadwal:
          type: object
          additionalProperties: false
          required: [tanggal, jam_mulai, jam_selesai, durasi]
          properties:
            tanggal: { $ref: '#/components/schemas/DateOnly' }
            jam_mulai: { $ref: '#/components/schemas/TimeHHmm' }
            jam_selesai: { $ref: '#/components/schemas/TimeHHmm' }
            durasi: { type: string, example: 3 Jam }
        rincian_pembayaran:
          type: object
          additionalProperties: false
          required: [tarif_kotor, diskon_promo, potongan, total_dibayar]
          properties:
            tarif_kotor: { $ref: '#/components/schemas/Money' }
            diskon_promo: { type: string, nullable: true }
            potongan: { $ref: '#/components/schemas/Money' }
            total_dibayar: { $ref: '#/components/schemas/Money' }
        status_reservasi: { $ref: '#/components/schemas/ReservationStatus' }
        qr_code_payload:
          type: string
          pattern: '^qr_[A-Za-z0-9_-]{20,}$'
          description: Token opaque/signed; tidak boleh berisi app key, JWT, atau PII.
    ETicketResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/ETicket' }
    ReservationMemberSummary:
      type: object
      additionalProperties: false
      required: [nama_member, telp]
      properties:
        id: { $ref: '#/components/schemas/PositiveId' }
        nama_member: { type: string }
        telp: { type: string }
    ReservationDetail:
      allOf:
        - { $ref: '#/components/schemas/Reservation' }
        - type: object
          required: [member, space]
          properties:
            member: { $ref: '#/components/schemas/ReservationMemberSummary' }
            space: { $ref: '#/components/schemas/SpaceSummary' }
    ReservationDetailResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/ReservationDetail' }
    ReservationStatusData:
      type: object
      additionalProperties: false
      required: [id, status, updated_at]
      properties:
        id: { $ref: '#/components/schemas/PositiveId' }
        status: { $ref: '#/components/schemas/ReservationStatus' }
        updated_at: { type: string, format: date-time }
    ReservationStatusResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/ReservationStatusData' }
...
    DeleteData:
      type: object
      additionalProperties: false
      required: [id, deleted]
      properties:
        id: { $ref: '#/components/schemas/PositiveId' }
        deleted: { type: boolean, enum: [true] }
    DeleteResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/DeleteData' }
...
    CreateDiscountDto:
      type: object
      additionalProperties: false
      required: [nama_diskon, persentase_diskon, tanggal_awal, tanggal_akhir]
      properties:
        nama_diskon: { type: string, minLength: 1, maxLength: 100, pattern: '^[A-Za-z0-9_-]+$', example: PROMOAGUSTUS }
        persentase_diskon: { type: integer, minimum: 1, maximum: 100, example: 20 }
        tanggal_awal: { type: string, format: date-time, example: '2026-08-01T00:00:00Z' }
        tanggal_akhir: { type: string, format: date-time, example: '2026-08-31T23:59:59Z' }
    UpdateDiscountDto:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nama_diskon: { type: string, minLength: 1, maxLength: 100, pattern: '^[A-Za-z0-9_-]+$' }
        persentase_diskon: { type: integer, minimum: 1, maximum: 100 }
        tanggal_awal: { type: string, format: date-time }
        tanggal_akhir: { type: string, format: date-time }
...